requirements.txt
# Core ML/DL frameworks
torch>=1.13.0
torchvision>=0.14.0

# MMSegmentation ecosystem
mmcv>=2.0.0
mmsegmentation>=1.0.0
mmengine>=0.7.0

# Data processing
numpy>=1.21.0
opencv-python>=4.6.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0

# Configuration
pyyaml>=6.0
tqdm>=4.64.0
